fix: update some files

+++ internal/db/migrations/00001_init.sql
@@ -3,13 +3,13 @@
 --+goose StatementBegin
 
 CREATE TABLE carts(
-                      id BIGSERIAL PRIMARY KEY,
+                      id INTEGER PRIMARY KEY,
                       created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
                       updated_at TIMESTAMP WITH TIME ZONE DEFAULT NOW()
 );
 
 CREATE TABLE cart_items (
-                            id BIGSERIAL PRIMARY KEY,
+                            id INTEGER PRIMARY KEY,
                             cart_id BIGINT NOT NULL, FOREIGN KEY(cart_id) REFERENCES carts(id) ON DELETE CASCADE,
                             product VARCHAR (255) NOT NULL,
                             price DECIMAL(10,2) NOT NULL CHECK(price>0),
